refactor: move command palette entries from useTerminals to App.tsx
The commands memo was a view concern (palette menu construction) living
inside a state management module. Move it to App.tsx — the composition
root that already extended it with app-level commands. This merges the
two-step commands()+allCommands pattern into a single unified memo.

useTerminals now exposes committedThemeName, setPreviewThemeName, and
handleSetTheme so App.tsx can build the theme sub-commands.
