+++ client/src/App.tsx
@@ -6,6 +6,7 @@ import {
   createEffect,
   createMemo,
   on,
+  batch,
   createResource,
   Show,
   For,
@@ -23,6 +24,7 @@ import MissionControl, { type MCMode } from "./MissionControl";
 import ModalDialog, { refocusTerminal } from "./ModalDialog";
 import Dialog from "@corvu/dialog";
 import { SHORTCUTS } from "./keyboard";
+import { availableThemes } from "./theme";
 
 import { client, wsStatus } from "./rpc";
 import { useTerminals } from "./useTerminals";
@@ -50,7 +52,9 @@ const App: Component = () => {
     getSubTerminalIds,
     reorderTerminals,
     mruOrder,
-    commands,
+    committedThemeName,
+    setPreviewThemeName,
+    handleSetTheme,
     randomTheme,
     setRandomTheme,
     scrollLock,
@@ -124,9 +128,48 @@ const App: Component = () => {
     setPaletteOpen(true);
   }
 
-  // Extend useTerminals commands with app-level commands (shortcuts help, about)
-  const allCommands = createMemo((): PaletteCommand[] => [
-    ...commands(),
+  // Command palette entries — all terminal + app-level commands in one place.
+  const commands = createMemo((): PaletteCommand[] => [
+    {
+      name: "Create new terminal",
+      keybind: SHORTCUTS.createTerminal.keybind,
+      onSelect: () => void handleCreate(),
+    },
+    ...(activeMeta()
+      ? [
+          {
+            name: "Create terminal in current directory",
+            keybind: SHORTCUTS.createTerminalInCwd.keybind,
+            onSelect: () => void handleCreate(activeMeta()!.cwd),
+          },
+        ]
+      : []),
+    ...(activeId() !== null
+      ? [
+          {
+            name: "Close terminal",
+            onSelect: () => void handleKill(activeId()!),
+          },
+          {
+            name: "Toggle sub-panel",
+            keybind: SHORTCUTS.toggleSubPanel.keybind,
+            onSelect: () => {
+              const id = activeId()!;
+              if (getSubTerminalIds(id).length === 0) {
+                void handleCreateSubTerminal(id, activeMeta()?.cwd);
+              } else {
+                subPanel.togglePanel(id);
+              }
+            },
+          },
+          {
+            name: "New sub-terminal",
+            keybind: SHORTCUTS.createSubTerminal.keybind,
+            onSelect: () =>
+              void handleCreateSubTerminal(activeId()!, activeMeta()?.cwd),
+          },
+        ]
+      : []),
     {
       name: "Mission Control",
       keybind: [
@@ -134,6 +177,40 @@ const App: Component = () => {
         SHORTCUTS.nextTerminalTab.keybind,
       ],
       onSelect: () => setMcMode({ mode: "browse" }),
+    },
+    ...(terminalIds().length > 0
+      ? [
+          {
+            name: "Switch terminal",
+            children: () =>
+              terminalIds().map((id, i) => ({
+                name: `Switch to terminal ${i + 1}`,
+                keybind:
+                  i < 9
+                    ? SHORTCUTS[
+                        `switchTo${(i + 1) as 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9}`
+                      ].keybind
+                    : undefined,
+                onSelect: () => setActiveId(id),
+              })),
+          },
+        ]
+      : []),
+    {
+      name: "Theme",
+      onCancel: () => setPreviewThemeName(undefined),
+      children: () =>
+        availableThemes
+          .filter((t) => t.name !== committedThemeName())
+          .map((t) => ({
+            name: t.name,
+            onHighlight: () => setPreviewThemeName(t.name),
+            onSelect: () =>
+              batch(() => {
+                setPreviewThemeName(undefined);
+                void handleSetTheme(t.name);
+              }),
+          })),
     },
     {
       name: "Keyboard shortcuts",
@@ -143,6 +220,20 @@ const App: Component = () => {
     {
       name: "About kolu",
       onSelect: () => setAboutOpen(true),
+    },
+    {
+      name: "Debug",
+      children: [
+        {
+          name: "Trigger server error",
+          onSelect: () =>
+            void client.terminal.resize({
+              id: "00000000-0000-0000-0000-000000000000",
+              cols: 1,
+              rows: 1,
+            }),
+        },
+      ],
     },
   ]);
 
@@ -184,7 +275,7 @@ const App: Component = () => {
         }}
       />
       <CommandPalette
-        commands={allCommands}
+        commands={commands}
         open={paletteOpen()}
         onOpenChange={handlePaletteOpenChange}
         initialGroup={paletteInitialGroup()}
